feat(details-page): move subcomponents to separate files
