+++ src/components/LayoutTemplates/UNSTABLE__DetailsPage1/index.js
@@ -3,6 +3,13 @@
  * @copyright IBM Security 2020
  */
 
+import UNSTABLE__DetailsPage1 from './UNSTABLE__DetailsPage1';
+import UNSTABLE__DetailsPage1Main from './UNSTABLE__DetailsPage1Main';
+import UNSTABLE__DetailsPage1Section from './UNSTABLE__DetailsPage1Section';
+import UNSTABLE__DetailsPage1SectionTerm from './UNSTABLE__DetailsPage1SectionTerm';
+import UNSTABLE__DetailsPage1SectionDefinition from './UNSTABLE__DetailsPage1SectionDefinition';
+import UNSTABLE__DetailsPage1Aside from './UNSTABLE__DetailsPage1Aside';
+
 export {
   UNSTABLE__DetailsPage1,
   UNSTABLE__DetailsPage1Main,
@@ -10,4 +17,4 @@ export {
   UNSTABLE__DetailsPage1SectionTerm,
   UNSTABLE__DetailsPage1SectionDefinition,
   UNSTABLE__DetailsPage1Aside,
-} from './UNSTABLE__DetailsPage1';
+};
