+++ src/index.js
@@ -241,7 +241,6 @@ export {
 export TabContent from 'carbon-components-react/lib/components/TabContent';
 
 export {
-  UNSTABLE__DetailsPage1,
   UNSTABLE__DetailsPage1Main,
   UNSTABLE__DetailsPage1Section,
   UNSTABLE__DetailsPage1SectionTerm,
